Api-Mobilesam-Task-Stage/requirements.txt
fastapi
uvicorn
Pillow
torch
torchvision
timm
opencv-python
git+https://github.com/dhkim2810/MobileSAM.git
matplotlib
numpy
